Upgrade to pgstac 0.4.5 (#61)
* Upgrade stac-fastapi / pgstac

Upgrades to an unreleased version of stac-fastapi to capture a pgstac
upgrade for item search improvements and cql2-text support. Also
upgrades fastapi version to match dependency requirements in stac-fastapi.

* Fix tests for new behavior of `limit`

When searching for ids, limit is no longer respected. You will get all
ids you requested. Two tests needed to be modified to account for this
behavior, to ensure that a paginated request was constructed, avoid
using ids and instead use a wide datetime to ensure a large response set
that can be limited.