Reregister redis lua scripts if missing (#69)
* Reregister redis lua scripts if missing

We've seen registered scripts be flushed from an active redis
connection. When that exception occurs, we now attempt to reregister the
scripts in a thread safe way, to avoid a sudden influx of
re-registrations.

* Avoid re-registering scripts on app startup

Service instances starting up may not need to register a script when
starting up. Consolidates register/re-register workflows.