Set low limit for number of items per tile
Some collections, due to use of skipfull or just the size of items
relative to a tile, could be hitting the hardcoded pgstac-titiler limit
for number of items to attempt to read into a tile. This lowers to a
strict default of 5, but allows collections to configure overrides
higher or lower if deemed necessary.