Include hostname in redis cache key (#98)
* Consolidate how dev depenencies are installed

Previous to this change pcstac and pctiler had
two methods for installing dev dependencies.
This caused different behavior between the two
test projects. For instance, one project auto
upgraded the pytest asyncio plugin, which
under a new version started printing warnings.

This users the setup.py method for installing dev
dependencies consistently and pins some test dependencies

* Use asyncio_mode = auto to quell warnings

* Fix dependencies in docker-compose

* Use hostname in caching key.

Previous to this change, requests coming in from
different hostnames were causing the cache to return
results for different hostnames if that service was
accessible through both - this is the case after a release,
when planetarycomputer-staging.msft.com and planetarycomputer.msft.com
both point to the same stack. This was causing cached
results to be returned with links for staging through
requests to the production endpoints. Added the hostname
to the cache key will ensure cached results are scoped
to the specific hostname.

* Update CHANGELOG