Timeout middleware  (#120)
* Exclude successful health check endpoints from logs

* Add timeout middleware

This middleware is intended to return a controlled message and error to
the user in cases where the request exceeds the timeout. Previously,
they might recieve a QueryCancelledError, a Service Unavailable, or a
GatewayTimeout depending on which mechanism (still sometimes opaque to
us) initiated the timeout. By ensuring the setting is less than the
gateway timeout, we should collect logs in all instances and provide a
meaningful error message to the user.  # Please enter the commit message
for your changes.

* Add support email in error message

* Changelog