HTML-escape inputs before templating (#155)
* Jinja filter

* Sanitize values before templating into map page

There was an XSS vulnerability allowing script tags to be injected via
the item or collection parameters, causing client side execution. Both
values are sanitized to avoid this vulnerability.

* Remove unused template

An unused template from the previous hard-coded mosaic era. Removing as
it does contain an XSS vulnerability were it to be reused.