Package publish permissions and ICU lib for dotnet functools (#189)
Sets a the write permission for the chart/func package publish jobs, which need to write to the gh-pages branch of the repo. This permission was previously set at the repo level, and may not be adjustable. Also adds a libicu dependency required for func-tools on ubuntu 20.04. This package has an interactive step, so much be installed in noninteractive mode.